refactor: Move the duplex, merge, and reverse tasks into FileawayCore This will allow us to easily use them in the macOS app in the future.
