+++ core/Sources/FileawayCore/Extensions/PDFDocument.swift
@@ -18,9 +18,27 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
+import Combine
+import Foundation
 import PDFKit
 
+enum PDFDocumentUtilitiesError: Error {
+    case failure
+}
+
 public extension PDFDocument {
+
+    static func open(url: URL) -> Future<PDFDocument, Error> {
+        return Future() { promise in
+            DispatchQueue.global(qos: .userInteractive).async {
+                guard let document = PDFDocument(url: url) else {
+                    promise(.failure(PDFDocumentUtilitiesError.failure))
+                    return
+                }
+                promise(.success(document))
+            }
+        }
+    }
 
     func dates() -> [Date] {
         let pageCount = self.pageCount
@@ -33,5 +51,31 @@ public extension PDFDocument {
         let dates = DateFinder.dateInstances(from: documentContent.string).map { $0.date }.uniqued()
         return dates
     }
+
+    func reverse() -> PDFDocument {
+        let documentCopy = self.copy() as! PDFDocument
+        for forwardIndex: Int in 0..<documentCopy.pageCount / 2 {
+            let backwardIndex = documentCopy.pageCount - forwardIndex - 1
+            documentCopy.exchangePage(at: forwardIndex, withPageAt: backwardIndex)
+        }
+        return documentCopy
+    }
+
+    func interleave(_ document: PDFDocument) -> PDFDocument {
+        let new = PDFDocument()
+        let documents: [PDFDocument] = [self.copy() as! PDFDocument, document.copy() as! PDFDocument]
+        while documents.map({ $0.pageCount }).reduce(0, +) > 0 {
+            for document in documents {
+                guard document.pageCount > 0 else {
+                    continue
+                }
+                let page = document.page(at: 0)?.copy() as! PDFPage
+                document.removePage(at: 0)
+                new.insert(page, at: new.pageCount)
+            }
+        }
+        return new
+    }
+
     
 }
