+++ core/Sources/FileawayCore/Tasks/MergeTask.swift
@@ -22,12 +22,16 @@ import Combine
 import Foundation
 import PDFKit
 
-class FakeDuplexTask {
+public class MergeTask {
 
     private var task: AnyCancellable?
 
-    func perform(url1: URL, url2: URL, output: URL, completion: @escaping (Result<Bool, Error>) -> Void) {
-        task = Publishers.Zip(PDFDocument.open(url: url1), PDFDocument.open(url: url2).map { $0.reverse() })
+    public init() {
+
+    }
+
+    public func merge(url1: URL, url2: URL, output: URL, completion: @escaping (Result<Bool, Error>) -> Void) {
+        task = Publishers.Zip(PDFDocument.open(url: url1), PDFDocument.open(url: url2))
             .map { $0.interleave($1) }
             .map { $0.write(to: output) }
             .receive(on: DispatchQueue.main)
@@ -41,3 +45,4 @@ class FakeDuplexTask {
     }
 
 }
+
