+++ core/Sources/FileawayCore/Tasks/ReverseTask.swift
@@ -22,14 +22,19 @@ import Combine
 import Foundation
 import PDFKit
 
-class MergeTask {
+public class ReverseTask {
 
     private var task: AnyCancellable?
 
-    func merge(url1: URL, url2: URL, output: URL, completion: @escaping (Result<Bool, Error>) -> Void) {
-        task = Publishers.Zip(PDFDocument.open(url: url1), PDFDocument.open(url: url2))
-            .map { $0.interleave($1) }
-            .map { $0.write(to: output) }
+    public init() {
+
+    }
+
+    public func reverse(url: URL, completion: @escaping (Result<Bool, Error>) -> Void) {
+        self.task = PDFDocument.open(url: url)
+            .map { $0.reverse() }
+            .map { $0.write(to: url) }
+            .eraseToAnyPublisher()
             .receive(on: DispatchQueue.main)
             .sink(receiveCompletion: { result in
                 if case .failure(let error) = result {
@@ -37,8 +42,7 @@ class MergeTask {
                 }
             }) { document in
                 completion(.success(true))
-            }
+        }
     }
 
 }
-
